ui-provider 분리
