+++ lib/src/screens/heart/heart_view_model.dart
@@ -1,41 +1,10 @@
 import 'package:flutter/material.dart';
 
-import '../../models/heart/heart_list_model.dart';
-import '../../models/heart/heart_service.dart';
-import '../../models/user/user_provider.dart';
-
 class HeartViewModel with ChangeNotifier {
   bool isChecked = false; //교환가능만 보기 버튼
 
-  int userId = UserProvider.userId;
-  bool isInitial = true;
-
-  HeartListModel? _heartListModel;
-  List<HeartListContentModel>? _heartList = [];
-  List<HeartListContentModel>? get heartList => _heartList;
-
-  void setInitialFalse() {
-    isInitial = false;
-    notifyListeners();
-  }
-
   void onButtonTapped() {
     isChecked = !isChecked;
-    getHeartList();
-    // heartListProvider.setShowAll();
-    notifyListeners();
-  }
-
-  List<HeartListContentModel>? getHeartList() {
-    fetchHeartList();
-    notifyListeners();
-    return _heartList;
-  }
-
-  fetchHeartList() async {
-    _heartListModel = await getHeartsByUser(userId, !isChecked);
-    _heartList = _heartListModel?.content;
-
     notifyListeners();
   }
 }
